Major overhaul to make this compatible with dtube's avalon chain
Right now this is only outputting data and allowing login with only 1 key.
It is using any key you use as it would be an active key on hive, but what you can really do depends on the key permissions (anyways in the future, it only outputs data right now).
Also, it doesn't check for key correctness, A.K.A. if the private key matches any public key on blockchain.